requirements.txt
fastapi==0.96.0
matplotlib==3.7.1
python-dotenv==1.0.0
python-multipart==0.0.6
rasterio==1.3.7
requests==2.31.0
uvicorn==0.22.0